Move footer page types to place where they are used
